REB3-8088. Fixed task creation with cached index.

+++ cvat/settings/base.py
@@ -464,7 +464,7 @@ RESTRICTIONS = {
 
 # http://www.grantjenks.com/docs/diskcache/tutorial.html#djangocache
 USE_CACHE = bool(int(os.getenv('USE_CACHE', 1)))
-CACHE_EXPIRE = os.getenv('CACHE_EXPIRE', 7 * 24 * 60 * 60)  # week in seconds
+CACHE_EXPIRE = float(os.getenv('CACHE_EXPIRE', 7 * 24 * 60 * 60))  # week in seconds
 
 CACHES = {
     'default': {
